+++ e2e/desktop/tests/page/index.ts
@@ -7,6 +7,7 @@ import { BuyAndSellPage } from "./buyAndSell.page";
 import { DelegateDrawer } from "./drawer/delegate.drawer";
 import { DelegateModal } from "./modal/delegate.modal";
 import { Drawer } from "../component/drawer.component";
+import { EarnPage } from "./earn.dashboard.page";
 import { Layout } from "../component/layout.component";
 import { LedgerSyncDrawer } from "./drawer/ledger.sync.drawer";
 import { LiveApp } from "./liveApp.page";
@@ -43,6 +44,7 @@ export class Application extends PageHolder {
   public delegate = new DelegateModal(this.page);
   public delegateDrawer = new DelegateDrawer(this.page);
   public drawer = new Drawer(this.page);
+  public earnDashboard = new EarnPage(this.page, this.electronApp);
   public layout = new Layout(this.page);
   public ledgerSync = new LedgerSyncDrawer(this.page);
   public liveApp = new LiveApp(this.page);