+++ libs/coin-modules/coin-evm/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ledgerhq/coin-evm
 
+## 2.22.5
+
+### Patch Changes
+
+- Updated dependencies [[`881dc73`](https://github.com/LedgerHQ/ledger-live/commit/881dc73e98bba95567fc2a4a6c54659d44f22957), [`5735489`](https://github.com/LedgerHQ/ledger-live/commit/5735489ddcee66110fc0cccc6bdd696876b8be4d)]:
+  - @ledgerhq/evm-tools@1.7.0
+  - @ledgerhq/cryptoassets@13.20.0
+  - @ledgerhq/live-env@2.11.0
+  - @ledgerhq/coin-framework@5.4.1
+  - @ledgerhq/domain-service@1.2.34
+  - @ledgerhq/live-network@2.0.12
+
+## 2.22.5-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`881dc73`](https://github.com/LedgerHQ/ledger-live/commit/881dc73e98bba95567fc2a4a6c54659d44f22957), [`5735489`](https://github.com/LedgerHQ/ledger-live/commit/5735489ddcee66110fc0cccc6bdd696876b8be4d)]:
+  - @ledgerhq/evm-tools@1.7.0-next.0
+  - @ledgerhq/cryptoassets@13.20.0-next.0
+  - @ledgerhq/live-env@2.11.0-next.0
+  - @ledgerhq/coin-framework@5.4.1-next.0
+  - @ledgerhq/domain-service@1.2.34-next.0
+  - @ledgerhq/live-network@2.0.12-next.0
+
 ## 2.22.4
 
 ### Patch Changes